# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/feliratkozasok": {
      "get": {
        "description": "Lekéri az összes feliratkozást (védett végpont).",
        "operationId": "FeliratkozasokController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Feliratkozások sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feliratkozas"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Összes feliratkozás lekérdezése",
        "tags": [
          "Feliratkozások"
        ]
      },
      "post": {
        "description": "Új feliratkozás létrehozása autentikált felhasználó számára.",
        "operationId": "FeliratkozasokController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeliratkozasDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feliratkozas sikeresen létrehozva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feliratkozas"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Új feliratkozás létrehozása",
        "tags": [
          "Feliratkozások"
        ]
      }
    },
    "/feliratkozasok/{id}": {
      "get": {
        "description": "Lekéri egy adott feliratkozás adatait az azonosító alapján.",
        "operationId": "FeliratkozasokController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feliratkozás sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feliratkozas"
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "404": {
            "description": "Feliratkozás nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Feliratkozás lekérdezése ID alapján",
        "tags": [
          "Feliratkozások"
        ]
      },
      "delete": {
        "description": "Meglévő feliratkozás törlése autentikált felhasználó számára.",
        "operationId": "FeliratkozasokController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feliratkozás sikeresen törölve"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "404": {
            "description": "Feliratkozás nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Feliratkozás törlése",
        "tags": [
          "Feliratkozások"
        ]
      }
    },
    "/feliratkozasok/szemely/{id}": {
      "get": {
        "description": "Lekéri egy adott személyhez tartozó összes eseményt.",
        "operationId": "FeliratkozasokController_szemelyEsemenyei",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Események sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feliratkozas"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "404": {
            "description": "Személy nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Személy eseményeinek lekérdezése",
        "tags": [
          "Feliratkozások"
        ]
      }
    },
    "/feliratkozasok/esemeny/{id}": {
      "get": {
        "description": "Lekéri egy adott eseményhez tartozó összes személyt.",
        "operationId": "FeliratkozasokController_esemenySzemelyei",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Személyek sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feliratkozas"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "404": {
            "description": "Esemény nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Esemény személyeinek lekérdezése",
        "tags": [
          "Feliratkozások"
        ]
      }
    },
    "/hirek": {
      "get": {
        "description": "Lekéri az összes hírt (publikus végpont).",
        "operationId": "HirekController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Hírek sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hir"
                  }
                }
              }
            }
          }
        },
        "summary": "Összes hír lekérdezése",
        "tags": [
          "Hírek"
        ]
      },
      "post": {
        "description": "Új hír létrehozása admin jogosultsággal.",
        "operationId": "HirekController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHirDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hír sikeresen létrehozva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hir"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Új hír létrehozása",
        "tags": [
          "Hírek"
        ]
      }
    },
    "/hirek/{id}": {
      "get": {
        "description": "Lekéri egy adott hír adatait az azonosító alapján (publikus végpont).",
        "operationId": "HirekController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hír sikeres lekérdezése",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hir"
                }
              }
            }
          },
          "404": {
            "description": "Hír nem található"
          }
        },
        "summary": "Hír lekérdezése ID alapján",
        "tags": [
          "Hírek"
        ]
      },
      "patch": {
        "description": "Meglévő hír módosítása admin jogosultsággal.",
        "operationId": "HirekController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHirDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hír sikeresen módosítva",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hir"
                }
              }
            }
          },
          "400": {
            "description": "Érvénytelen adatok"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          },
          "404": {
            "description": "Hír nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Hír módosítása",
        "tags": [
          "Hírek"
        ]
      },
      "delete": {
        "description": "Meglévő hír törlése admin jogosultsággal.",
        "operationId": "HirekController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hír sikeresen törölve"
          },
          "401": {
            "description": "Érvénytelen vagy hiányzó JWT token"
          },
          "403": {
            "description": "Nincs admin jogosultság"
          },
          "404": {
            "description": "Hír nem található"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Hír törlése",
        "tags": [
          "Hírek"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS keretrendszerrel készített REST API",
    "description": "REST API személyek, címek, események, feliratkozások, hírek és települések kezelésére",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Esemeny": {
        "type": "object",
        "properties": {
          "esemeny_id": {
            "type": "number",
            "description": "Az esemény egyedi azonosítója",
            "example": 1
          },
          "esemeny_nev": {
            "type": "string",
            "description": "Az esemény neve",
            "example": "Éves konferencia"
          },
          "esemeny_leiras": {
            "type": "string",
            "description": "Az esemény leírása",
            "example": "Éves szakmai konferencia a technológiai újításokról"
          },
          "esemeny_helyszin": {
            "type": "string",
            "description": "Az esemény helyszíne",
            "example": "Budapest, Kongresszusi Központ"
          },
          "esemeny_idopont": {
            "format": "date-time",
            "type": "string",
            "description": "Az esemény időpontja",
            "example": "2025-12-01T10:00:00Z"
          },
          "esemeny_max_fo": {
            "type": "number",
            "description": "Az esemény maximális létszáma",
            "example": 100
          },
          "esemeny_feliratkozasok": {
            "description": "Az eseményhez tartozó feliratkozások",
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "esemeny_id",
          "esemeny_nev",
          "esemeny_leiras",
          "esemeny_helyszin",
          "esemeny_idopont",
          "esemeny_max_fo",
          "esemeny_feliratkozasok"
        ]
      },
      "Feliratkozas": {
        "type": "object",
        "properties": {
          "feliratkozas_id": {
            "type": "number",
            "description": "A feliratkozás egyedi azonosítója",
            "example": 1
          },
          "feliratkozas_szemely_id": {
            "type": "number",
            "description": "A feliratkozó személy azonosítója",
            "example": 1
          },
          "feliratkozas_esemeny_id": {
            "type": "number",
            "description": "Az esemény azonosítója",
            "example": 1
          },
          "feliratkozas_esemeny": {
            "description": "A feliratkozás eseménye",
            "allOf": [
              {
                "$ref": "#/components/schemas/Esemeny"
              }
            ]
          }
        },
        "required": [
          "feliratkozas_id",
          "feliratkozas_szemely_id",
          "feliratkozas_esemeny_id",
          "feliratkozas_esemeny"
        ]
      },
...
      "CreateFeliratkozasDto": {
        "type": "object",
        "properties": {
          "feliratkozas_szemely_id": {
            "type": "number",
            "description": "A feliratkozó személy azonosítója",
            "example": 1
          },
          "feliratkozas_esemeny_id": {
            "type": "number",
            "description": "Az esemény azonosítója",
            "example": 1
          }
        },
        "required": [
          "feliratkozas_szemely_id",
          "feliratkozas_esemeny_id"
        ]
      },
      "Hir": {
        "type": "object",
        "properties": {
          "hir_id": {
            "type": "number",
            "description": "A hír egyedi azonosítója",
            "example": 1
          },
          "hir_cim": {
            "type": "string",
            "description": "A hír címe",
            "example": "Új esemény bejelentése"
          },
          "hir_tartalom": {
            "type": "string",
            "description": "A hír tartalma",
            "example": "Új esemény kerül megrendezésre decemberben."
          },
          "hir_datum": {
            "type": "object",
            "description": "A hír dátuma (opcionális)",
            "example": "2025-06-28"
          }
        },
        "required": [
          "hir_id",
          "hir_cim",
          "hir_tartalom"
        ]
      },
      "CreateHirDto": {
        "type": "object",
        "properties": {
          "hir_cim": {
            "type": "string",
            "description": "A hír címe",
            "example": "Új esemény bejelentése"
          },
          "hir_tartalom": {
            "type": "string",
            "description": "A hír tartalma",
            "example": "Új esemény kerül megrendezésre decemberben."
          },
          "hir_datum": {
            "type": "object",
            "description": "A hír dátuma (opcionális)",
            "example": "2025-06-28"
          }
        },
        "required": [
          "hir_cim",
          "hir_tartalom"
        ]
      },
      "UpdateHirDto": {
        "type": "object",
        "properties": {
          "hir_cim": {
            "type": "string",
            "description": "A hír címe (opcionális)",
            "example": "Új esemény bejelentése"
          },
          "hir_tartalom": {
            "type": "string",
            "description": "A hír tartalma (opcionális)",
            "example": "Új esemény kerül megrendezésre decemberben."
          },
          "hir_datum": {
            "type": "object",
            "description": "A hír dátuma (opcionális)",
            "example": "2025-06-28"
          }
        }
      },
...
    }
  }
}